Rename packages (#20)
