+++ packages/vsf-storyblok-extension/package.json
@@ -1,8 +1,8 @@
 {
-  "name": "@kodbruket/vsf-storyblok-sync",
+  "name": "@kodbruket/vsf-storyblok-extension",
   "version": "0.0.1",
-  "description": "Vue Storefront Storyblok consumer",
-  "main": "index.ts",
+  "description": "Vue Storefront Storyblok extension",
+  "main": "index.js",
   "repository": "https://github.com/kodbruket/vsf-storyblok-sync",
   "author": "Max Malm",
   "license": "MIT",
