+++ packages/vsf-storyblok-module/hooks/beforeRegistration.ts
