--- packages/vsf-storyblok-sync/pages/Storyblok.vue
@@ -1,17 +1,0 @@
-<template>
-  <div>
-    <sb-render v-if="story" :item="story.content" />
-  </div>
-</template>
-
-<script>
-import StoryblokMixin from 'src/modules/storyblok-sync/components/StoryblokMixin'
-import Render from 'src/modules/storyblok-sync/components/Render.vue'
-export default {
-  name: 'StoryblokPage',
-  mixins: [StoryblokMixin],
-  components: {
-    sbRender: Render
-  }
-}
-</script>
